refactor: Small reorg of parser error
- Move the prqlc-wide `Error` our of `parser` path
- Collapse a small multifile `mod.rs` into a single file (think we should do more of this)
