+++ prqlc/prqlc-parser/src/parser/mod.rs
@@ -1,4 +1,11 @@
+use chumsky::{prelude::*, Stream};
+
+use crate::error::Error;
+use crate::lexer::lr;
+use crate::span::Span;
+
 mod common;
+pub(crate) mod error;
 mod expr;
 mod interpolation;
 pub mod pr;
@@ -6,12 +13,6 @@ pub(crate) mod stmt;
 #[cfg(test)]
 mod test;
 mod types;
-
-use chumsky::{prelude::*, Stream};
-
-use crate::error::Error;
-use crate::lexer::lr;
-use crate::span::Span;
 
 pub fn parse_lr_to_pr(
     source: &str,
