+++ prqlc/prqlc-parser/src/parser/stmt.rs
@@ -6,8 +6,8 @@ use semver::VersionReq;
 
 use super::common::{ctrl, ident_part, into_stmt, keyword, new_line};
 use super::expr::{expr, expr_call, ident, pipeline};
-use crate::error::parse_error::PError;
 use crate::lexer::lr::{Literal, TokenKind};
+use crate::parser::error::PError;
 use crate::parser::pr::*;
 use crate::parser::types::type_expr;
